Bug fixes and automatically resizing window
if you create a tile map that would not fit in the current window, it is
resized to accomadate the tile map (but not bigger than the screen
size).
Fixed bug when using the rectangle tool and releasing over the tile set
panel would cause you to select a different tile.