Added horizontal scrollable tile maps
the window no longer automatically resizes when you create a tile map
that is too big, instead (only in the horizontal direction for now) a
scroll bar appears. You can use it to scroll through the tile map.

Bug:
the drawing and selection box on the tile map don't work properly yet.