Changed automatic tile set opening
Instead of searching the current directory and the .exe directory for
the tilesheet, the program now creates its own "tilesheets" directory
and whenever you import a tile sheet it is copied into that directory.
Then, when you open a tile map and the proper tile set is not opened
yet, the program searches the "tilesheets" directory for the tile set.